+++ validator/validations/shapes/validations/shape_points_coordinates_consistent_validation.go
@@ -19,8 +19,8 @@ type shapePointsCoordinatesConsistentPoint struct {
 
 func buildShapeFromPointsCoordinatesConsistentPoint(point shapePointsCoordinatesConsistentPoint) *types.Shape {
 	return &types.Shape{
-		ShapePtLat: lib.Ptr(float32(point.lat)),
-		ShapePtLon: lib.Ptr(float32(point.lon)),
+		ShapePtLat: lib.Ptr(float64(point.lat)),
+		ShapePtLon: lib.Ptr(float64(point.lon)),
 	}
 }
 
@@ -70,8 +70,8 @@ func ShapePointsCoordinatesConsistentValidation(shapes []types.Shape, rules *typ
 			id:       *shape.ShapeId,
 			row:      i,
 			sequence: *shape.ShapePtSequence,
-			lat:      float64(*shape.ShapePtLat),
-			lon:      float64(*shape.ShapePtLon),
+			lat:      *shape.ShapePtLat,
+			lon:      *shape.ShapePtLon,
 		})
 	}
